--- simulation/arrays/__init__.py
@@ -1,2 +1,0 @@
-from simulation.arrays.base_array import BaseArray
-from simulation.arrays.basic_array import BasicArray
