--- simulation/battery/__init__.py
@@ -1,2 +1,0 @@
-from simulation.battery.base_battery import BaseBattery
-from simulation.battery.basic_battery import BasicBattery
