+++ simulation/cmd/run_simulation.py
@@ -5,9 +5,7 @@ import json
 import sys
 import csv
 
-from simulation.main import Simulation, SimulationReturnType
-from simulation.optimization.bayesian import BayesianOptimization
-from simulation.optimization.random_opt import RandomOptimization
+from simulation.model.Simulation import Simulation, SimulationReturnType
 from simulation.utils.InputBounds import InputBounds
 from simulation.config import config_directory
 from simulation.utils.SimulationBuilder import SimulationBuilder
