--- simulation/main/__init__.py
@@ -1,2 +1,0 @@
-from simulation.main.Simulation import Simulation, SimulationReturnType
-from simulation.main.Model import Model
