+++ simulation/model/car/battery/base_battery.py
@@ -1,4 +1,3 @@
-from simulation.common import BatteryEmptyError
 from simulation.common import Storage
