+++ simulation/model/car/lvs/basic_lvs.py
@@ -1,4 +1,4 @@
-from simulation.lvs.base_lvs import BaseLVS
+from simulation.model.car.lvs.base_lvs import BaseLVS
 from simulation.common import DayBreak
