+++ simulation/model/car/motor/__init__.py
@@ -1,0 +1,2 @@
+from simulation.model.car.motor.base_motor import BaseMotor
+from simulation.model.car.motor.basic_motor import BasicMotor
