+++ simulation/model/car/motor/base_motor.py
