--- simulation/motor/__init__.py
@@ -1,2 +1,0 @@
-from simulation.motor.base_motor import BaseMotor
-from simulation.motor.basic_motor import BasicMotor
