--- simulation/regen/__init__.py
@@ -1,2 +1,0 @@
-from simulation.regen.base_regen import BaseRegen
-from simulation.regen.basic_regen import BasicRegen
